+++ packages/rspack/src/container/ModuleFederationPlugin.ts
@@ -3,7 +3,7 @@ import { Compiler } from "../Compiler";
 import { type ModuleFederationPluginV1Options } from "./ModuleFederationPluginV1";
 import { parseOptions } from "./options";
 import { isValidate } from "../util/validate";
-import { FederationRuntimePlugin } from "./FederationRuntimePlugin";
+import { ModuleFederationRuntimePlugin } from "./ModuleFederationRuntimePlugin";
 
 export interface ModuleFederationPluginOptions
 	extends Omit<ModuleFederationPluginV1Options, "enhanced"> {
@@ -30,7 +30,7 @@ export class ModuleFederationPlugin {
 				{ name: undefined }
 			).apply(compiler);
 		});
-		new FederationRuntimePlugin().apply(compiler);
+		new ModuleFederationRuntimePlugin().apply(compiler);
 		new webpack.container.ModuleFederationPluginV1({
 			...this._options,
 			enhanced: true
