+++ packages/rspack/src/container/ModuleFederationRuntimePlugin.ts
@@ -1,7 +1,7 @@
 import { BuiltinPluginName } from "@rspack/binding";
 import { create } from "../builtin-plugin/base";
 
-export const FederationRuntimePlugin = create(
-	BuiltinPluginName.FederationRuntimePlugin,
+export const ModuleFederationRuntimePlugin = create(
+	BuiltinPluginName.ModuleFederationRuntimePlugin,
 	() => {}
 );
