+++ packages/rspack/tests/configCases/container-1-5/federation-instance-in-runtime-plugin/bootstrap.js
@@ -1,0 +1,5 @@
+export function test(it) {
+  it("should able to access federation object on __webpack_require__", () => {
+    expect(__webpack_require__.federation.test).toBe(1)
+  })
+}
