+++ packages/rspack/tests/configCases/container-1-5/federation-instance-in-runtime-plugin/index.js
@@ -1,0 +1,3 @@
+it("should import the correct modules", () => {
+	return import("./bootstrap").then(({ test }) => test(it))
+});
