+++ packages/rspack/tests/configCases/container-1-5/federation-instance-in-runtime-plugin/plugin.js
@@ -1,0 +1,3 @@
+export default function () {
+  __webpack_require__.federation.test = 1
+}
