+++ packages/rspack/tests/configCases/container-1-5/federation-instance-in-runtime-plugin/webpack.config.js
@@ -1,0 +1,10 @@
+const { ModuleFederationPlugin } = require("@rspack/core").container;
+
+/** @type {import("@rspack/core").Configuration} */
+module.exports = {
+	plugins: [
+		new ModuleFederationPlugin({
+			runtimePlugins: ["./plugin.js"]
+		})
+	]
+};
